requirements.txt
langchain
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
transformers
torch
python-dotenv
pypdf

fastapi
uvicorn
streamlit
requests
pydantic
